Create model group

+++ database/migrations/2020_04_02_075207_create_offices_table.php
@@ -15,7 +15,7 @@ class CreateOfficesTable extends Migration
     {
         Schema::create('offices', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->string('name');
             $table->string('post_number');
             $table->string('address');
